refactor : view Lotto Money 분리
