+++ src/main/java/lotto/view/money/MoneyOutput.java
@@ -1,0 +1,12 @@
+package lotto.view.money;
+
+public class MoneyOutput {
+
+    public static void purchaseMoney() {
+        System.out.println("구입금액을 입력해 주세요.");
+    }
+
+    public static void moneyInputError() {
+        System.out.println("구매금액 입력이 잘못됐습니다 다시 입력해주세요.");
+    }
+}
